Eradicate CONTENT_LENGTH_UNKNOWN wherever possible
It seems that if we set this option, in order to omit the
Content-Length http header and in order to send unknown
amounts of data even after we send headers, that request will
become REALLY slow. This is a worthwhile tradeoff in only one
instance, which is for /setup. On this page, we are sending
a lot of data, too much for one string (sending the entire page
seems to overwhelm the ESP). But we also don't know the length
of all the data, because we are scanning for networks. So,
there are better ways, but the easiest is to use this. Good
thing we don't really care how fast this page loads, as it
won't be used day to day, just once for setup.

In every other instance, if we set up a String with all the
content we want to send, and then send it in one shot with
send(), the loading time is MUCH better. Also, as a bonus
a Content-Length header is calculated and sent by the ESP
web server library.